整理 GnuFilter 类
